+++ cognite_toolkit/_cdf_tk/builders/__init__.py
@@ -1,6 +1,8 @@
 from pathlib import Path
 
-from ._base import Builder, FileBuilder, FunctionBuilder
+from ._base import Builder
+from ._file import FileBuilder
+from ._function import FunctionBuilder
 
 
 def create_builder(
